Split API runtime state into focused modules
